+++ src/test/mocks/utils/levels-loader.js
@@ -1,0 +1,5 @@
+//var properties = require('../properties-mock');
+
+module.exports = {
+  loadLevelsJson: function() {}
+};