+++ src/test/states/load-spec.js
@@ -8,9 +8,10 @@ var mocks = require('mocks');
 
 var stubs = {
   '../properties': require('../mocks/properties-mock'),
-  '../data/level-manager': require('../mocks/data/level-manager')
+  '../utils/levels-loader': require('../mocks/utils/levels-loader')
 };
 
+//var levelsLoader = require('../utils/levels-loader');
 
 var proxyquire = require('proxyquireify')(require);
 var state = proxyquire('../../app/states/load', stubs);
@@ -39,19 +40,9 @@ describe("Phaser load state tests", function() {
       expect(game.load.atlas).to.have.been.calledWith('dpad', 'assets/images/virtualjoystick/skins/dpad.png', 'assets/images/virtualjoystick/skins/dpad.json');
     });
 
-    it('combined levels and actors texture atlas should be loaded', function() {
-      state.preload();
-      expect(game.load.atlas).to.have.been.calledWith('combined', 'assets/atlas/combined.png', 'assets/atlas/combined.json');
-    });
-
     it('player phyics data should be loaded', function() {
       state.preload();
       expect(game.load.physics).to.have.been.calledWith('playerPhysics', 'assets/physics/player.json');
-    });
-
-    it('all level map physics data should be loaded', function() {
-      state.preload();
-      expect(game.load.physics).to.have.been.calledWith('mapPhysics-map', 'assets/levels/level_6.json');
     });
 
   });